// src/models/tablero.model.ts
export interface Tablero {
  1: string;
  2: string;
  3: string;
  4: string;
  5: string;
  6: string;
  7: string;
  8: string;
  9: string;
}

export interface ActiveStyles {
  box_1: object;
  box_2: object;
  box_3: object;
  box_4: object;
  box_5: object;
  box_6: object;
  box_7: object;
  box_8: object;
  box_9: object;
}